"valores"

+++ app/cart/templatetags/cart_tags.py
@@ -5,7 +5,7 @@ register = template.Library()
 
 @register.filter()
 def multiply(value, arg):
-    return float(value) * arg
+    return int(value) * arg
 
 
 @register.filter()
